pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ptp-instance-tester"
version = "0.1.0"
description = "A tool to test PTP hardware clock support on AWS EC2 instance types"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "boto3>=1.28.0",
    "paramiko>=3.3.0",
    "click>=8.1.0",
    "pyyaml>=6.0",
]

[project.scripts]
ptp-tester = "ptp_tester.cli:main"

[tool.setuptools.packages.find]
where = ["src"]
